adding reusable carousel pattern tested and works
