+++ src/app/pages/projects/page.tsx
@@ -1,83 +1,26 @@
 'use client';
-import { Card, CardContent } from '@/components/ui/card';
 
-import {
-  Carousel,
-  CarouselContent,
-  CarouselItem,
-  CarouselNext,
-  CarouselPrevious,
-} from '@/components/ui/carroussel';
-
-import * as React from 'react';
-import Autoplay from 'embla-carousel-autoplay';
+import CarouselPattern from '@/components/patterns/carousel-pattern';
 
 export default function ProjectsPage() {
-  const plugin = React.useRef(
-    Autoplay({ delay: 2000, stopOnInteraction: true })
-  );
+  const items = [
+    {
+      text: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Deleniti laboriosam incidunt iure facilis enim velit perferendis, corporis soluta sunt aperiam quibusdam laborum. Laborum aperiam temporibus voluptate quis velit facilis blanditiis!',
+      mainImagePath: '/assets/weather-app.png',
+      githubPath: 'https://github.com/alexandre-bouchard/weather-app',
+      videoPath: 'https://www.youtube.com/watch?v=l9eZ3g_V72k',
+    },
+    {
+      text: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Deleniti laboriosam incidunt iure facilis enim velit perferendis, corporis soluta sunt aperiam quibusdam laborum. Laborum aperiam temporibus voluptate quis velit facilis blanditiis!',
+      mainImagePath: '/assets/weather-app.png',
+      githubPath: 'https://github.com/alexandre-bouchard/weather-app',
+      videoPath: 'https://www.youtube.com/watch?v=l9eZ3g_V72k',
+    },
+  ];
 
   return (
-    <div className="grid place-items-center w-full overflow-hidden">
-      <Carousel
-        plugins={[plugin.current]}
-        className="w-11/12 sm:w-9/12 lg:w-8/12"
-        onMouseEnter={plugin.current.stop}
-        onMouseLeave={plugin.current.reset}
-      >
-        <CarouselContent>
-          {Array.from({ length: 5 }).map((_, index) => (
-            <CarouselItem key={index}>
-              <div className="p-2 sm:p-4">
-                <Card>
-                  <CardContent className="flex flex-col items-center justify-center w-full space-y-4">
-                    <div className="view_project">
-                      <img
-                        src={`/assets/weather-app.gif`}
-                        alt="Little girl constructing something"
-                        className="w-full h-auto max-h-64 sm:max-h-72 lg:max-h-96 rounded-lg shadow-lg object-cover"
-                      />
-                    </div>
-
-                    <div className="presentation_content text-center px-4">
-                      <p className="text-xs sm:text-sm text-gray-600 dark:text-gray-300">
-                        A weather data API developed with NodeJS, NestJS,
-                        Typescript, PostgreSQL, TypeORM, Docker.
-                      </p>
-                    </div>
-
-                    <div className="presentation__buttons-link flex space-x-4">
-                      <button className="focus:outline-none">
-                        <a
-                          href="https://github.com/upALX/Orpen-Weather"
-                          target="_blank"
-                        >
-                          <img
-                            className="w-6 h-6 sm:w-8 sm:h-8 transition-transform hover:scale-110"
-                            src="/assets/go-code.png"
-                            alt="Code link"
-                          />
-                        </a>
-                      </button>
-                      <button className="focus:outline-none">
-                        <a href="https://youtu.be/nFREOIxeqig" target="_blank">
-                          <img
-                            className="w-6 h-6 sm:w-8 sm:h-8 transition-transform hover:scale-110"
-                            src="/assets/go-project.png"
-                            alt="Project demo link"
-                          />
-                        </a>
-                      </button>
-                    </div>
-                  </CardContent>
-                </Card>
-              </div>
-            </CarouselItem>
-          ))}
-        </CarouselContent>
-        <CarouselPrevious />
-        <CarouselNext />
-      </Carousel>
+    <div className="grid place-items-center  min-h-screen w-full overflow-hidden">
+      <CarouselPattern items={items} />
     </div>
   );
 }
