+++ src/components/types/carousel-pattern-types.ts
@@ -1,0 +1,10 @@
+export type ItemsType = {
+  text: string;
+  mainImagePath: string;
+  videoPath: string;
+  githubPath: string;
+};
+
+export type CarouselPatternProps = {
+  items: ItemsType[];
+};
